Fix shares combination and rewrite overall code
There was a bug for combining shares where the prime number was being
choose from the size of a random secret, thus obtaining the wrong
prime number to combine. It wasn't detected by tests because it worked
for certain combinations. Another bug was to not check for the number
of shares to be at least equal to the threshold.

Fix the generation of integers for the polynomial in SecretShare so
that the number of bits of those integers are the required and not
more.

Change Secret random method to be an object method and receiving the
number of bits as as a parameter. Additionally verify that the
generated secret is lower than the corresponding prime.

Move primes to its own module, with its tests, and add more (bigger)
primes.

Rewrite code to base Secret and Share value on integer numbers
instead of bytes to avoid several bugs affecting the code.

Add missing tests for Share and Secret classes, and fix base methods:
add to_hex and remove __hex__ and __oct__.